refactor: eslint format
