+++ apps/cms-admin/src/components/layout/AppHeader.tsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from 'react'
 import { Link, useNavigate } from 'react-router-dom'
 import { Dropdown, Icon, Menu } from 'semantic-ui-react'
-import { appMenus } from 'src/appMenu'
+import { appMenus } from 'src/app-menu'
 import Notification from 'src/components/Notification'
 import useAuthorizationContext from 'src/contexts/authorization/authorization.context'
-import useModuleName from 'src/hooks/useModuleName'
+import useModuleName from 'src/hooks/use-module-name'
 import { routeMap } from 'src/route'
 
 import './AppHeader.scss'
