+++ apps/cms-admin/src/contexts/authorization/use-authorization.test.ts
@@ -1,7 +1,7 @@
 import { MemoryRouter } from 'react-router-dom'
 import { act, renderHook, waitFor } from '@testing-library/react'
 import type { Mock } from 'vitest'
-import useAuthorization from 'src/contexts/authorization/useAuthorization'
+import useAuthorization from 'src/contexts/authorization/use-authorization'
 import { service } from 'src/services'
 import StorageUtil from 'src/utils/storage.util'
 import type { AuthRo } from '../../services/api'
