+++ apps/cms-admin/src/index.tsx
@@ -4,7 +4,7 @@ import * as ReactDOM from 'react-dom/client'
 import App from 'src/App'
 import { AuthorizationProvider } from 'src/contexts/authorization/authorization.context'
 import { ToastProvider } from 'src/contexts/toast/toast.context'
-import * as serviceWorker from 'src/serviceWorker'
+import * as serviceWorker from 'src/service-worker'
 
 import 'semantic-ui-css/semantic.min.css'
 import 'src/assets/css/index.scss'
