+++ apps/cms-admin/src/pages/content/article/ArticleCreatePage.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom'
 import { Header, Icon, Menu, Segment } from 'semantic-ui-react'
 import FormRenderer from 'src/components/form/FormRenderer'
 import useToast from 'src/contexts/toast/toast.context'
-import { articleForm, articleFormConfig } from 'src/pages/content/article/articleForm.config'
+import { articleForm, articleFormConfig } from 'src/pages/content/article/article-form.config'
 import { service, useSubmit } from 'src/services'
 import type { CreateArticleDto } from 'src/services/api'
