+++ apps/cms-admin/src/services/hooks/use-retrieve-detail.ts
