+++ apps/cms-admin/src/services/hooks/use-submit.test.ts
@@ -1,6 +1,6 @@
 import React from 'react'
 import { act, renderHook } from '@testing-library/react'
-import { isFormError, useSubmit } from './useSubmit'
+import { isFormError, useSubmit } from './use-submit'
 
 describe('# isFormError', () => {
   it('should return true when error has data message and status code is 422', () => {
