+++ apps/cms-admin/src/services/index.ts
@@ -1,8 +1,8 @@
 import CONFIG from 'src/config'
 import { Api } from './api'
 
-export * from './hooks/useSubmit'
-export * from './hooks/useRetrieveList'
+export * from './hooks/use-submit'
+export * from './hooks/use-retrieve-list'
 
 export const service = new Api<string>({
   baseURL: CONFIG.API_URL,
