+++ apps/cms-admin/tsconfig.json
@@ -35,6 +35,6 @@
     "src"
   ],
   "exclude": [
-    "src/setupTests.ts"
+    "src/setup-tests.ts"
   ]
 }
