+++ apps/cms-admin/tsconfig.node.json
@@ -8,6 +8,6 @@
   "include": [
     "./vite.config.ts",
     "./vitest.config.ts",
-    "./src/setupTests.ts"
+    "./src/setup-tests.ts"
   ]
 }
