+++ apps/cms-page-builder/src/modules/FullWidthCarouselModule.vue
@@ -15,7 +15,7 @@
 
 <script setup lang="ts">
 import { computed } from 'vue'
-import { currentSection } from 'src/stores/pageBuilder'
+import { currentSection } from 'src/stores/page-builder'
 import Button from '../components/Button.vue'
 import Carousel from '../components/Carousel.vue'
