+++ apps/cms-page-builder/src/utils/__test__/to-style.spec.ts
@@ -1,5 +1,5 @@
-import type { StyleObject } from '../toStyle'
-import { toStyle } from '../toStyle'
+import type { StyleObject } from '../to-style'
+import { toStyle } from '../to-style'
 
 describe('# toStyle', () => {
   it('should convert style object to kebab-case from camelCase', () => {
