+++ apps/cms-page-builder/src/utils/get-scrollbar-width.ts
