+++ apps/cms-page-builder/src/utils/index.ts
@@ -1,6 +1,6 @@
 export * from './debounce'
 export * from './throttle'
-export * from './generateId'
-export * from './toStyle'
+export * from './generate-id'
+export * from './to-style'
 export * from './pick'
-export * from './getScrollbarWidth'
+export * from './get-scrollbar-width'
