+++ package.json
@@ -17,7 +17,7 @@
   },
   "devDependencies": {
     "@codecov/vite-plugin": "0.0.1-beta.10",
-    "@mutoe/eslint-config": "2.27.3-2",
+    "@mutoe/eslint-config": "3.1.0-1",
     "@types/node": "^20",
     "@vitest/coverage-v8": "^2.0.5",
     "@vitest/ui": "^2.0.5",
